Removed haaukins files made for refactoring purpose
