+++ svcs/guacamole/keylogger.go
@@ -7,7 +7,6 @@ package guacamole
 import (
 	"bytes"
 	"github.com/aau-network-security/haaukins/logging"
-	"github.com/aau-network-security/haaukins/store"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"io"
@@ -179,7 +178,8 @@ func NewKeyLogger(logger *zerolog.Logger) (KeyLogger, error) {
 }
 
 type KeyLoggerPool interface {
-	GetLogger(t store.Team) (KeyLogger, error)
+	//todo commented stufff
+	//GetLogger(t store.Team) (KeyLogger, error)
 	io.Closer
 }
 
@@ -187,13 +187,14 @@ type keyLoggerPool struct {
 	logpool logging.Pool
 }
 
-func (klp *keyLoggerPool) GetLogger(t store.Team) (KeyLogger, error) {
-	logger, err := klp.logpool.GetLogger(t.Id)
-	if err != nil {
-		return nil, err
-	}
-	return NewKeyLogger(logger)
-}
+//todo uncomment this
+//func (klp *keyLoggerPool) GetLogger(t store.Team) (KeyLogger, error) {
+//	logger, err := klp.logpool.GetLogger(t.Id)
+//	if err != nil {
+//		return nil, err
+//	}
+//	return NewKeyLogger(logger)
+//}
 
 func (klp *keyLoggerPool) Close() error {
 	return klp.logpool.Close()
